requirements.txt
# FREE VERSION - No paid APIs required!
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
groq>=0.4.0
requests>=2.31.0
pydantic>=2.0.0
qrcode>=7.4.2
pillow>=10.0.0
pymongo>=4.6.0
